install finalizers for the reader and Writer structs
so that if, for some reason, the calling code fails to Close() the
reader or Writer, we do not leak the z.zstrm's C malloc allocations.

Note that it is safe to call (*reader).Close() and (*Writer).Close()
multiple times: they both set z.err to io.EOF, which then causes
the later calls to short circuit.